fix: navbar should be fixed to the bottom
Due to Firefox Android scrolling issues with its own address bar disappearing on scroll, I thought the simplest and correct way of doing it was not working.
Actually it works in Chrome Android and Safari on iPhone.

A new known issue with Firefox Android is introduced: lagging scrolling bahavior of the navbar.

Ref: https://bugzilla.mozilla.org/show_bug.cgi?id=1880375